fix: Change aid type from int to long (#1775)
* fix: Set Aid type to long

* fix: Update cid data type

+++ VocaDbModel/Service/VideoServices/VideoServiceBilibili.cs
@@ -104,24 +104,24 @@ public class VideoServiceBilibili : VideoService
 public class BiliMetadata
 {
 	[DataMember]
-	public int Aid { get; init; }
+	public long Aid { get; init; }
 	[DataMember]
 	public string Bvid { get; init; } = default!;
 	[DataMember]
-	public int Cid { get; init; }
+	public long Cid { get; init; }
 }
 
 class BilibiliResponse
 {
-	public int Code { get; init; } = default!;
+	public long Code { get; init; } = default!;
 	public BilibiliResponseData? Data { get; init; } = default;
 }
 
 class BilibiliResponseData
 {
-	public int Aid { get; init; }
+	public long Aid { get; init; }
 	public string Bvid { get; init; } = default!;
-	public int Cid { get; init; }
+	public long Cid { get; init; }
 	[JsonConverter(typeof(UnixDateTimeConverter))]
 	public DateTime? PubDate { get; init; }
 	public int Duration { get; init; }
